tested batch manager on aegir
